// src/lib/strapi/types/api/fiche-solution.ts
export enum TypeSolution {
  Bleue = 'bleue',
  Verte = 'verte',
  Grise = 'grise',
  Douce = 'douce',}
export enum PorteeBaisseTemperature {
  Air = 'air',
  Surface = 'surface',
  Interieur = 'interieur',}
export enum CoutEntretienUnite {
  MetreCarre = 'metreCarre',
  Lineaire = 'lineaire',
  MetreCube = 'metreCube',
  Unite = 'unite',
  MegaWattHeure = 'megaWattHeure',
  KiloWatt = 'kiloWatt',}
export enum CoutUnite {
  MetreCarre = 'metreCarre',
  MegaWattHeure = 'megaWattHeure',}

export interface FicheSolution {
  id: number;
  attributes: {
    createdAt: Date;    updatedAt: Date;    publishedAt?: Date;    titre: string;
    slug: string;
    image_principale?: { data: Media };
    solution_retour_experiences: { data: SolutionRetourExperience[] };
    cout_minimum?: number;
    cout_maximum?: number;
    type_solution?: TypeSolution;
    description_courte: string;
    description: any;
    baisse_temperature?: number;
    contexte_titre: string;
    contexte_description: any;
    rafraichissement_attendu_description?: any;
    logo_partenaire?: { data: Media };
    cout_minimum_entretien?: number;
    cout_maximum_entretien?: number;
    etapes_mise_en_oeuvre: EtapeMiseEnOeuvre[];
    etapes_entretien: EtapeEntretien[];
    point_vigilance?: any;
    lien_aide_territoire?: string;
    lien_fond_vert?: string;
    types_espace?: any;
    etapes_diagnostic: EtapeDiagnostic[];
    aides_regionales: AideRegionale[];
    rank?: number;
    materiaux?: { data: Materiau[] };
    fiches_solutions_complementaires: { data: FicheSolution[] };
    cobenefices?: { data: Cobenefice[] };
    oups: Oups[];
    cout_entretien_description?: any;
    delai_travaux_minimum?: number;
    delai_travaux_maximum?: number;
    portee_baisse_temperature?: PorteeBaisseTemperature;
    cout_entretien_unite?: CoutEntretienUnite;
    en_savoir_plus?: any;
    libelle_avantage_solution?: string;
    credits?: any;
    cout_unite?: CoutUnite;
    description_estimation?: any;
    aides_territoires_mots_cles?: string;
  };
}